+++ contracts/PriceOracle/interfaces/XSushiExchangeRateInterface.sol
@@ -1,5 +1,5 @@
 pragma solidity ^0.5.16;
 
 interface XSushiExchangeRateInterface {
-    function getExchangeRate() external view returns (uint);
+    function getExchangeRate() external view returns (uint256);
 }
